Add Biome-Makeover, Structory, and Terralith structures.
Add the structures from Biome-Makeover, Structory, and Terralith which
have minecraft:lectern and/or minecraft:lectern in their nbt data.
It is worth nothing that some of the
structures (e.g. biomemakeover:mansion) will not have
replacements for bookcases generate since that mod does not appear to
use the net.minecraft.structure.processor.StructureProcessor when
generating the mansion structures.